+++ core/src/Type/Definition/Search/StepResult/CypherPathSubsetStepResult.ts
@@ -1,0 +1,10 @@
+import { Uuid } from '../../Uuid.js';
+
+interface CypherPathSubsetStepResult {
+  paths: {
+    nodeIds: Uuid[];
+    relationIds: Uuid[];
+  };
+}
+
+export { CypherPathSubsetStepResult };