+++ core/src/Type/Definition/Search/StepResult/ElementHydrationStepResult.ts
@@ -1,0 +1,6 @@
+import { Node } from '../../Node.js';
+import { Relation } from '../../Relation.js';
+
+type ElementHydrationStepResult = [Node, Relation][];
+
+export { ElementHydrationStepResult };