+++ core/src/Type/Definition/Search/index.ts
@@ -1,0 +1,2 @@
+export * as Step from './Step/index.js';
+export * as StepResult from './StepResult/index.js';